feat: better color order
